requirements.txt
openai==0.28.1
tiktoken
langchain
pinecone-client
